LayoutTests/platform/ios-simulator/editing/pasteboard/paste-text-with-style-3-expected.txt

This tests copying and pasting text does not result in nested style elements such as b, em, and font.
To manually test, copy "hello world" below and paste it in the box below. WebKit should not nest b, i, u, and font.

Before copy:
| <i>
|   "hello"

After paste:
| <i>
|   style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|   "hello<#selection-caret>"
| <font>
|   size="5"
|   <u>
|     <br>
